added homepage v0.05 and studyoptions v3.0 into pages folder, studyoptions are now responsive, redesigned footer
